Remove dead code to fix a lint warning
Obsolete SDK_INT Version Check
------------------------------
This check flags version checks that are not necessary,
because the minSdkVersion (or surrounding known API level)
is already at least as high as the version checked for.

The mindSdkVersion in templates/AndroidManifest.xml is 16 for Qt 5.12
and androiddeployqt also ensures that the minSdkVersion is not below 16.

Change-Id: Ifdb6f773dccdb75c35119df833c976b7b76b3356
Reviewed-by: BogDan Vatra <bogdan@kdab.com>