SQL: add proper support to build QMYSQL with MariaDB client libraries
The QMYSQL plugin can also be build with the MariaDB client libraries
since they are source compatible. But the MariaDB libraries could not be
found on windows because the library name differs. Therefore add the
correct library names and update the documentation to make clear that
MariaDB is supported through the QMYSQL plugin.

[ChangeLog][Sql][QMYSQL] The QMYSQL plugin can now be build with the
MariaDB C connector libs on Windows.

Change-Id: Id99f8be96c4179fd2321b3e61c90bb300c53bb82
Reviewed-by: Marius Kittler <mariuskittler@gmx.de>
Reviewed-by: Andy Shaw <andy.shaw@qt.io>