Freetype: Don't embolden bold fonts
If a font does not set the "bold" flag in its OS/2 table, we check
the weight from the same table to determine whether it is bold or
if we have to embolden it synthetically. But the actual definition of
bold in OS/2 is 700 (which is also what QFont::Bold is documented
to correspond to in qfont.h).

The result was that we would embolden fonts with bold weight if
the bold flag was not set. An example of such a font was the
CJK JP family of Noto Sans.

[ChangeLog][Text] Fixed a problem where certain bold fonts would be
synthetically emboldened by Qt when using the Freetype font engine.

Fixes: QTBUG-80866
Change-Id: I2133d9c44a9e19c0f5f216a649ec64388245d34f
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>